cloudpebble/requirements.txt
Django==1.6.2
django-registration==1.0
amqp==1.4.9
celery==3.1.23
django-celery==3.1.17
flup==1.0.3.dev-20110405
psycopg2==2.8.3
South==1.0.2
qrcode==4.0.4
Pillow==2.9.0
pygithub==1.14.2
python-social-auth==0.1.23
redis==2.10.5
boto==2.39.0
gevent==1.1
psycogreen==1.0
requests==2.7.0
oauth2client==1.3
django-bower==5.0.4
mock==2.0.0
# This is used for Mailchimp integration
mailchimp==2.0.9

# These are used for Heroku hosting setup.
dj-static==0.0.6
gunicorn==19.3.0
dj-database-url==0.3.0
newrelic==2.50.0.39

# These are in the SDK requirements.txt and included here to save on Heroku deploy time.
freetype-py==1.0
sh==1.08
pypng==0.0.17

# Additional requirements
anyjson==0.3.3
billiard==3.3.0.23
docopt==0.4.0
greenlet==0.4.9
httplib2==0.9.2
kombu==3.0.35
oauthlib==1.0.3
pyasn1==0.1.7
pyasn1-modules==0.0.5
python-openid==2.2.5
pytz==2016.3
requests-oauthlib==0.6.1
rsa==3.1.4
six==1.10.0
static3==0.6.1

python-memcached
